forms_rest/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-rest-swagger==2.2.0
psycopg2-binary==2.9.9
djoser==2.2.2
parameterized==0.9.0